Add files via upload
Hey Gail.  These are for a micro Blog assignment.  I have the Create User and Create Post.

You can see the Main Menu in "Main" and that's where I'm sort of struggling.

We're supposed to let them: 

Create new User (I assume loop through the Create User class and store it in an array list. I'm struggling with how to get the loop calls to add to a list. After they make a new user it should return them to the Main Menu again)

Become an existing user.  My thought process is to have it print out a list of Created users and they can type in which one they want to be, or the users can be assigned a number via a list. then loop back to menu.

Create a post, it's supposed to print the last post by the current user and prompt for the user to type another post.  Since it's required that the users last post be printed I assume save in an array.  (again.  The struggle.  I know array list would be best but still.)

Print all posts (Pull from the array of posts and print them all)  (I have already created a "Create Post" class.

Print all users (Same as print all posts